LPS-66921 Move FolderUtil too as it is used only by the data handlers and it was in the original lar package
