+++ modules/apps/collaboration/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/exportimport/data/handler/FolderUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.document.library.internal.lar;
+package com.liferay.document.library.internal.exportimport.data.handler;
 
 import com.liferay.document.library.kernel.exception.NoSuchFolderException;
 import com.liferay.document.library.kernel.model.DLFolder;
